+++ collector/dvvp/analyze/src/analyzer_ge.cpp
@@ -11,6 +11,7 @@
 #include "data_struct.h"
 #include "errno/error_code.h"
 #include "message/codec.h"
+#include "msprof_dlog.h"
 #include "proto/profiler.pb.h"
 #include "transport/hash_data.h"
 
@@ -38,7 +39,7 @@ void AnalyzerGe::Parse(SHARED_PTR_ALIA<analysis::dvvp::proto::FileChunkReq> mess
         return;
     }
     if (message->tag().find("id_map_info") != std::string::npos) {
-        totalBytes_ += message->chunksizeinbytes();
+        totalBytes_ += static_cast<uint64_t>(message->chunksizeinbytes());
         if (message->chunksizeinbytes() < GE_ID_MAP_SIZE) {
             MSPROF_LOGE("id_map_info is incomplete data");
             return;
@@ -47,7 +48,7 @@ void AnalyzerGe::Parse(SHARED_PTR_ALIA<analysis::dvvp::proto::FileChunkReq> mess
         return;
     }
     if (message->tag().find("task_desc_info") != std::string::npos) {
-        totalBytes_ += message->chunksizeinbytes();
+        totalBytes_ += static_cast<uint64_t>(message->chunksizeinbytes());
         ParseTaskDesc(message->chunk().c_str(), message->chunksizeinbytes());
         return;
     }
